# backend/api/openapi/swagger.yaml
definitions:
...
  sico-backend_internal_transport_http_dto_rbac_common.User:
    properties:
      alias:
        type: string
      createdAt:
        type: integer
      description:
        type: string
      email:
        type: string
      iconUri:
        type: string
      id:
        type: integer
      phone:
        type: string
      rawIconUri:
        type: string
      roles:
        items:
          type: string
        type: array
      status:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.UserStatus'
      tenant:
        type: string
      updatedAt:
        type: integer
      username:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_common.UserStatus:
    enum:
    - 0
    - 1
    - 2
    format: int32
    type: integer
    x-enum-varnames:
    - UserStatus_USER_STATUS_INACTIVE
    - UserStatus_USER_STATUS_ACTIVE
    - UserStatus_USER_STATUS_FROZEN
...
  sico-backend_internal_transport_http_dto_rbac_role.Role:
    properties:
      code:
        type: string
      createdAt:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        enum:
        - 0
        - 1
        type: integer
      updatedAt:
        type: integer
    type: object
...
  sico-backend_internal_transport_http_dto_rbac_user_role.AssignUserRoleRequest:
    properties:
      roleId:
        type: integer
      userId:
        type: integer
    required:
    - roleId
    - userId
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.AssignUserRoleResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.ListUserRolesResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.ListUserRolesResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.ListUserRolesResponseData:
    properties:
      hasNext:
        type: boolean
      roles:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_role.Role'
        type: array
      total:
        type: integer
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.ListUsersByRoleResponse:
    properties:
      code:
        type: integer
      data:
        $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.ListUsersByRoleResponseData'
      msg:
        type: string
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.ListUsersByRoleResponseData:
    properties:
      hasNext:
        type: boolean
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_common.User'
        type: array
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.RemoveUserRoleRequest:
    properties:
      roleId:
        type: integer
      userId:
        type: integer
    required:
    - roleId
    - userId
    type: object
  sico-backend_internal_transport_http_dto_rbac_user_role.RemoveUserRoleResponse:
    properties:
      code:
        type: integer
      msg:
        type: string
    type: object
...
info:
  contact: {}
  description: Sico AI Agent Platform API
  title: Sico API
  version: "1.0"
paths:
...
  /api/sico/rbac/role_users:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - in: query
        name: roleId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.ListUsersByRoleResponse'
      security:
      - BearerAuth: []
      summary: List Users By Role
      tags:
      - RBAC
...
  /api/sico/rbac/user_role:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Remove User Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.RemoveUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.RemoveUserRoleResponse'
      security:
      - BearerAuth: []
      summary: Remove User Role
      tags:
      - RBAC
    post:
      consumes:
      - application/json
      parameters:
      - description: Assign User Role
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.AssignUserRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.AssignUserRoleResponse'
      security:
      - BearerAuth: []
      summary: Assign User Role
      tags:
      - RBAC
  /api/sico/rbac/user_roles:
    get:
      consumes:
      - application/json
      parameters:
      - default: 1
        in: query
        minimum: 1
        name: page
        type: integer
      - default: 10
        in: query
        maximum: 100
        minimum: 1
        name: pageSize
        type: integer
      - in: query
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/sico-backend_internal_transport_http_dto_rbac_user_role.ListUserRolesResponse'
      security:
      - BearerAuth: []
      summary: List User Roles
      tags:
      - RBAC
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"